Load `zstd` with pkg-config
As Nix aims to be the single source of dependencies, all Cargo dependencies
are bundled and provided as _read only_ vendored sources [1].
Without the (optional) `pkg-config` feature, `zstd-sys` is configured to
build the `zstd` library from source as part of its own build process [2].
However, `zstd-sys` is located in a read only directory,
so when its build script tries to copy  files [3], this step fails with a
`Permission Denied` error.

relevant excerpt from the build log:

     >   thread 'main' panicked at 'called `Result::unwrap()` on an `Err` value: Os { code: 13, kind: PermissionDenied, message: "Permission denied" }', /private/tmp/nix-build-hello-rust-0.0.0.drv-0/cargo-vendor-dir/zstd-sys-2.0.4+zstd.1.5.2/build.rs:211:58

The solution (for this crate) is to enable the `pkg-config` feature,
which will employ `pkg-config` to link the rust wrapper to a pre-built `zstd`.
Since now nix needs to supply `zstd`,
we have to add it as an import in the nix package definition.

[1]: https://doc.rust-lang.org/cargo/reference/source-replacement.html
[2]: https://github.com/gyscos/zstd-rs/blob/90d080e2d296d6dd09336a041958340e118fe436/zstd-safe/zstd-sys/build.rs#L234
[3]: https://github.com/gyscos/zstd-rs/blob/90d080e2d296d6dd09336a041958340e118fe436/zstd-safe/zstd-sys/build.rs#L205-L210